基本完成

+++ src/main/java/com/carl/yimai/service/impl/AddressServiceImpl.java
@@ -5,6 +5,7 @@ import com.carl.yimai.mapper.YmUserAddrMapper;
 import com.carl.yimai.po.YmUserAddr;
 import com.carl.yimai.po.YmUserAddrExample;
 import com.carl.yimai.service.AddressService;
+import com.carl.yimai.service.OrderService;
 import com.carl.yimai.web.utils.Result;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -28,6 +29,9 @@ public class AddressServiceImpl implements AddressService {
     @Resource(name = "ymUserAddrMapper")
     private YmUserAddrMapper addrMapper;
 
+    @Resource(name = "orderService")
+    private OrderService orderService;
+
     /**
      * 获取用户的地址信息
      * @param id 用户的id
@@ -39,6 +43,17 @@ public class AddressServiceImpl implements AddressService {
         example.createCriteria().andUidEqualTo(id);
         List<YmUserAddr> addrs = addrMapper.selectByExample(example);
         return Result.ok(addrs);
+    }
+
+    @Override
+    public Result getAddressById(String addressId) {
+        YmUserAddr addr = addrMapper.selectByPrimaryKey(addressId);
+
+        if (null == addr) {
+            return Result.error("没有对应的地址的信息");
+        }
+
+        return Result.ok(addr.getAddress());
     }
 
     @Override
@@ -60,8 +75,30 @@ public class AddressServiceImpl implements AddressService {
             return Result.error("没有相关的地址信息");
         }
 
+        Result result = orderService.checkAddress(addr.getId());
+
+        if (!result.isStatus()) {
+            return result;
+        }
+
         //执行删除
         addrMapper.deleteByPrimaryKey(id);
+
+        return Result.ok();
+    }
+
+    /**
+     * 查询用户的地址的信息是否正确
+     * @param userId
+     * @param addrId
+     * @return
+     */
+    public Result checkAddress(String userId,String addrId){
+        YmUserAddr addr = addrMapper.selectByPrimaryKey(addrId);
+
+        if (null == addr || !userId.equals(addr.getUid())) {
+            return Result.error("没有对应的地址信息");
+        }
 
         return Result.ok();
     }
